Refactor worker (#1001)
* move code update logic away from deciding avatar actions

* remove legacy function from handle turn

* if statement not needed

* fixed print collector error

* minor changes

* remove globals from runner

* clean up imports

* fix import

* import fix++

* minor fixes

* fix incorrect logic in worker

* add CodeUpdater to tests

* fix imports

* re-add strange logic until it can be looked into more

* make aimmo update properly again

* make code handle SyntaxErrors better

* improve type error handling and fix tests

* Merge branch 'development' into refactor_worker

* bump version

* Merge branch 'development' into refactor_worker

* remove logger
