+++ aimmo-game-worker/avatar_runner.py
@@ -1,8 +1,6 @@
 from __future__ import print_function
 
 import contextlib
-import imp
-import inspect
 import logging
 import re
 import sys
@@ -21,41 +19,12 @@ from simulation.action import Action, WaitAction
 from user_exceptions import InvalidActionException
 
 LOGGER = logging.getLogger(__name__)
-
-
-def add_actions_to_globals():
-    action_classes = filter(lambda x: x[1].__module__ == "simulation.action", inspect.getmembers(avatar_action, inspect.isclass))
-
-    for action_class in action_classes:
-        restricted_globals[action_class[0]] = action_class[1]
-
-
-_getattr_ = safer_getattr
-_setattr_ = guarded_setattr
-_write_ = full_write_guard
-__metaclass__ = type
-
-# Sets up restricted coding environment for the user
 log_manager = LogManager()
-restricted_globals = dict(__builtins__=safe_builtins)
-
-restricted_globals['_getattr_'] = _getattr_
-restricted_globals['_setattr_'] = _setattr_
-restricted_globals['_getiter_'] = list
-restricted_globals['_print_'] = log_manager.get_print_collector()
-restricted_globals['_write_'] = _write_
-restricted_globals['__metaclass__'] = __metaclass__
-restricted_globals['__name__'] = "Avatar"
-
-# Adds AI:MMO specific modules to the user's environment
-add_actions_to_globals()
-restricted_globals['direction'] = direction
-restricted_globals['random'] = utility_builtins['random']
 
 
-# Temporarily switches stdout and stderr to stringIO objects or variable
 @contextlib.contextmanager
 def capture_output(stdout=None, stderr=None):
+    """Temporarily switches stdout and stderr to stringIO objects or variable."""
     old_out = sys.stdout
     old_err = sys.stderr
 
@@ -72,59 +41,22 @@ def capture_output(stdout=None, stderr=None):
 
 
 class AvatarRunner(object):
-    def __init__(self, avatar=None, auto_update=True):
+    def __init__(self, avatar=None, auto_update=True, code_updater=None):
         self.avatar = avatar
         self.auto_update = auto_update
-        self.avatar_source_code = None
-        self.update_successful = False
-
-    def _avatar_src_changed(self, new_avatar_code):
-        return new_avatar_code != self.avatar_source_code
-
-    def _get_new_avatar(self, src_code):
-        self.avatar_source_code = src_code
-        module = imp.new_module('avatar')  # Create a temporary module to execute the src_code in
-        module.__dict__.update(restricted_globals)
-
-        try:
-            byte_code = compile_restricted(src_code, filename='<inline-code>', mode='exec')
-            exec(byte_code, restricted_globals)
-        except SyntaxWarning as w:
-            pass
-
-        module.__dict__['Avatar'] = restricted_globals['Avatar']
-        return module.Avatar()
-
-    def _update_avatar(self, src_code):
-        """
-        We update the avatar object if any of the following are true:
-        1. We don't have an avatar object yet, so self.avatar is None
-        2. The new source code we have been given is different
-        3. If the previous attempt to create an avatar object failed (i.e. _get_new_avatar threw an exception)
-        The last condition is necessary because if _get_new_avatar fails the avatar object will not have
-        been updated, meaning that self.avatar will actually be for the last correct code
-        """
-
-        if self._should_update(src_code):
-            try:
-                self.avatar = self._get_new_avatar(src_code)
-                self.update_successful = True
-            except SyntaxError as e:
-                self.update_successful = False
-                print(e)
-            except Exception as e:
-                self.update_successful = False
-                raise e
-
-    def _should_update(self, src_code):
-        return (self.avatar is None or self.auto_update and self._avatar_src_changed(src_code) or
-                not self.update_successful)
+        self.code_updater = code_updater
+        self.code_updater.globals['_print_'] = log_manager.get_print_collector()
 
     def process_avatar_turn(self, world_map, avatar_state, src_code):
-        avatar_updated = self._avatar_src_changed(src_code)
-
+        avatar_updated = False
         with capture_output() as output:
-            action = self.run_users_code(world_map, avatar_state, src_code)
+            if self.code_updater.should_update(self.avatar, self.auto_update, src_code):
+                self.avatar, avatar_updated = self.code_updater.update_avatar(src_code)
+
+            if self.avatar:
+                action = self.run_users_code(world_map, avatar_state, src_code)
+            else:
+                action = WaitAction().serialise()
 
         stdout, stderr = output
         output_log = stdout.getvalue()
@@ -135,7 +67,6 @@ class AvatarRunner(object):
 
     def run_users_code(self, world_map, avatar_state, src_code):
         try:
-            self._update_avatar(src_code)
             action = self.decide_action(world_map, avatar_state)
             self.print_logs()
 
@@ -143,7 +74,10 @@ class AvatarRunner(object):
             self.print_logs()
             print(e)
             action = WaitAction().serialise()
-
+        except AttributeError as e:
+            self.print_logs()
+            print(e)
+            action = WaitAction().serialise()
         except Exception as e:
             self.print_logs()
             user_traceback = self.get_only_user_traceback()
@@ -158,16 +92,14 @@ class AvatarRunner(object):
 
     def decide_action(self, world_map, avatar_state):
         try:
-            try:
-                action = self.avatar.handle_turn(world_map, avatar_state)
-            except AttributeError:
-                action = self.avatar.next_turn(world_map, avatar_state)
-
+            action = self.avatar.next_turn(world_map, avatar_state)
             if not isinstance(action, Action):
                 raise InvalidActionException(action)
+
             return action.serialise()
         except TypeError as e:
-            raise InvalidActionException(None)
+            print(e)
+            return WaitAction().serialise()
 
     @staticmethod
     def get_only_user_traceback():
