+++ aimmo-game/service.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+
 import ast
 import asyncio
 import json
